Detect symbol group headers by looking for flags.
DDTORD 1462 says that group headers have "flags clear", while symbols
must have either SYLCL or SYGBL set. Looking at the code in DDT that
reads the symbol table (ALDS/SLOD), that seems to be how it detects
group headers too. So we don't need to guess whether a file is old or
new format any more; an old file just doesn't have any group headers.

This works on all the files in the samples/ directory, and on the
ancient binaries (including some from STINK) in the 1973 copy of Muddle.