Keep track of the original ordering of symbols explicitly.
If you have two symbols with the same address, the original code tried
to keep them in the order they were originally encountered, so the
earlier one was preferred. However, it assumed that qsort was a stable
sort, which isn't true on all platforms.

Keep track of the ordering explicitly in the symbol structure, and sort
by it as a last resort.