Add support for reading Muddle fast save files.
A save file contains the parts of the interpreter's memory other than the
interpreter code itself, including impure low memory and pages of pure
Muddle code/data that are stored or not according to Muddle's page map.

The locations of the variables necessary to do this vary between Muddle
versions and platforms, so the loader populates the symbol table based
on the version number in the file. It also includes some register and
variable definitions that are useful for understanding compiled Muddle
code.