Add word format for octal numbers, one word per line.
The parser tries to be somewhat accomodating; lines not recognized are
simply skipped.  Whitespace first on the line is ignored.  Then a
stretch of 12 octal digits is expected.  The next character must not
be an octal digit.  The rest of the line is ignored.

In particular, this is useful for binary files from saildart.org.