Add word format for Saildart.org files in UTF-8 format.
NUL characters are stripped out entirely.  The range 001-037 is mapped
to Unicode symbols corresponding to the SAIL character set.  The ASCII
standard characters ^ _ ~ } have different encodings, and old 1963
ASCII arrows are in place.

CRLF line endings are converted to a single LF.  It's unclear how
stray CR and LF characters should be handled; for now they are passed
through.