Update Makefile
Use $(MAKE) for recursive make invocation.   This respects the flags passed to the top level make as well as invoking the correct binary if the default is overridden on the command line.
Fixes a n ITS build failure on FreeBSD, which used `gmake MAKE=gmake EMULATOR=...` to recursively build everything using gmake instead of BSD make.